conversion script based on a config.json input file
